DateTime optimization
There's two kinds of optimization in this commit.

*	Common UNIX epoch object to make efficient and easy to read.
*	Use `Ticks' rather than `Seconds' in floating point values.

+++ CoreTweet/Internal/Converters.cs
@@ -119,7 +119,7 @@ namespace CoreTweet.Core
                     else
                         return new DateTimeOffset(((DateTime)reader.Value).ToUniversalTime(), TimeSpan.Zero);
                 case JsonToken.Integer:
-                    return InternalUtils.GetUnixTime((double)(long)reader.Value);
+                    return InternalUtils.GetUnixTime((long)reader.Value);
                 
                 case JsonToken.Null:
                     return DateTimeOffset.Now;
